requirements.txt
asgiref==3.2.10
Django==3.1.2
pytz==2020.1
sqlparse==0.3.1
confuse==1.3.0
psycopg2==2.8.6
djangorestframework==3.12.1
selenium==3.141.0
loguru==0.5.3
fake_useragent
bs4==0.0.1
requests==2.24.0
lxml==4.5.2
celery==5.0.0
redis==3.5.3
gevent==20.9.0
django-widget-tweaks==1.4.8
